Removed module EncounterModifier in favour of an Event
